Derive all the things
Extend the list of default extensions to include the full suite of deriving extensions. We want making types to be cheap and for types to be usable they often need to derive helpful typeclasses. By turning on the full suite of derivation extensions we make deriving cheaper and thus new types cheaper.