Specify naming guide for request/response types.
NOTE: This was an attempt to codify what we do most-commonly at this
time, not an attempt at finding an objectively good guide. If folks want
to propose an objectively good guide, have at it.

We are currently in violation of this guide because we're inconsistent
in casing the `Method` suffix, but these types are very easy to identify
and rename, so I plan to resolve things as soon as this is accepted.