Add DataKinds to default-extensions
This commit has our Guide match reality, as this was already enabled in
all packages anyway.

DataKinds is relatively innocuous and, while we consider it an _Isolate_
generally so as not to build things on it, we do _use_ quite a few thing
that require it be enabled. For example:

- Anything using a type-level symbol (Tagged, Closed, Envelope etc)
- Any Graphula tests with type signatures (i.e. GraphulaContext m '[])

Therefore, it makes sense in default-extensions.