Remove avoid MonadFail
The problem with `MonadFail` was specifically the `fail` method of
`Monad`. Because it was coupled with `Monad` is resulting in many
horrible instances, like the `Either` instance of `error` that is
mentioned here. After the Monad Fail Proposal `fail` has been pulled out
into its own class `MonadFail` that is useful and has limited instances
that aren't foot guns.